+++ app-server/src/interfaces/config.interface.ts
@@ -1,0 +1,5 @@
+export interface IConfig {
+    BING_API?: string;
+    PORT?: string
+    NODE_ENV?: string
+}