+++ app-web/src/api/backgroundApi.tsx
@@ -1,0 +1,25 @@
+import axios from "axios";
+import config from "../config/config";
+
+const instance = axios.create({
+    baseURL: config.APP_API,
+    headers: {
+        "Content-Type": "application/json"
+    }
+})
+
+export default {
+    getBackgroundImage: () => new Promise((resolve, reject) => {
+        instance.get('/background')
+            .then(response => {
+                if (response.status !== 200) {
+                    reject(new Error('Unable to get image from server'))
+                }
+                resolve(response.data)
+            })
+            .catch(error => {
+                console.error('getBackgroundImage error ', error)
+                reject(error)
+            })
+    })
+}