+++ app-web/src/index.tsx
@@ -3,14 +3,14 @@ import ReactDOM from 'react-dom/client'
 
 import Home from './pages/Home'
 import './styles/global.css'
-import { BingImageProvider } from './stores/HomeBackgroundProvider';
+import { BackgroundImageProvider } from './stores/HomeBackgroundProvider';
 
 const root = ReactDOM.createRoot(document.querySelector('#root') as Element);
 
 root.render(
     <React.StrictMode>
-        <BingImageProvider>
+        <BackgroundImageProvider>
             <Home />
-        </BingImageProvider>
+        </BackgroundImageProvider>
     </React.StrictMode>
 );