+++ app-web/src/interfaces/config.interface.ts
@@ -1,5 +1,5 @@
 export interface IConfig {
-    NODE_ENV: string;
     APP_NAME: string;
     BING_API: string;
+    APP_API: string;
 }