+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2024-06-21
+
+### Added
+- bing api request serverside
+- stores and context to get background image
+
 ## [0.5.0] - 2024-06-21
 
 ### Added